Split reconstruction code from _api.py into _reconstruction.py
Move _peak_table_from_list, _count_active_channels, _flatten_event_row,
_accumulate_peak_counts, and compute_peak_data_from_eventlist to a new
rsciio/tofwerk/_reconstruction.py module. Update __init__.py to import
compute_peak_data_from_eventlist from _reconstruction directly.
