+++ rsciio/tofwerk/__init__.py
@@ -1,4 +1,5 @@
-from ._api import available_signals, compute_peak_data_from_eventlist, file_reader
+from ._api import available_signals, file_reader
+from ._reconstruction import compute_peak_data_from_eventlist
 
 __all__ = [
     "available_signals",
