+++ rsciio/tofwerk/_reconstruction.py
@@ -1,0 +1,352 @@
+# -*- coding: utf-8 -*-
+# Copyright 2007-2026 The HyperSpy developers
+#
+# This file is part of RosettaSciIO.
+#
+# RosettaSciIO is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# RosettaSciIO is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with RosettaSciIO. If not, see <https://www.gnu.org/licenses/#GPL>.
+
+"""
+EventList reconstruction for Tofwerk TofDAQ HDF5 files.
+
+This module contains the peak-integration reconstruction functions that convert
+raw TDC EventList data into a 4-D ``(depth, y, x, m/z)`` peak-counts array,
+replicating the processing performed by the Tofwerk proprietary software.
+
+Public API
+----------
+compute_peak_data_from_eventlist
+    Reconstruct the 4-D peak-integrated array from a raw EventList.
+"""
+
+import logging
+import re
+
+import numpy as np
+
+from rsciio.utils._decorator import jit_ifnumba
+
+_logger = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# PeakTable helpers
+# ---------------------------------------------------------------------------
+
+
+def _peak_table_from_list(peak_table_list):
+    """
+    Convert a peak table list of dicts back to arrays suitable for integration.
+
+    Parameters
+    ----------
+    peak_table_list : list of dict
+        As returned by :func:`_peak_table_to_list`.
+
+    Returns
+    -------
+    masses : numpy.ndarray, float64
+    peak_low : numpy.ndarray, float64
+    peak_high : numpy.ndarray, float64
+    """
+    masses = np.array([p["mass"] for p in peak_table_list], dtype=float)
+    peak_low = np.array(
+        [p["lower_integration_limit"] for p in peak_table_list], dtype=float
+    )
+    peak_high = np.array(
+        [p["upper_integration_limit"] for p in peak_table_list], dtype=float
+    )
+    return masses, peak_low, peak_high
+
+
+# ---------------------------------------------------------------------------
+# EventList / peak data reconstruction helpers
+# ---------------------------------------------------------------------------
+
+
+def _count_active_channels(ini):
+    """
+    Return the number of active TDC recording channels from the INI config string.
+
+    TofDAQ records events from multiple ADC/TDC channels simultaneously
+    (e.g. Ch1 and Ch3 in a typical fibTOF setup).  Each active channel
+    contributes an independent copy of every ion event to the EventList,
+    so the raw event count per pixel is multiplied by this factor.
+
+    Parameters
+    ----------
+    ini : str
+        Contents of the ``Configuration File Contents`` root attribute.
+    """
+    matches = re.findall(r"Ch\dRecord=(\d)", ini)
+    return max(sum(int(m) for m in matches), 1)
+
+
+def _flatten_event_row(row):
+    """
+    Pack a row of variable-length event arrays into a flat contiguous buffer.
+
+    Parameters
+    ----------
+    row : array-like of arrays
+        Object array of shape ``(nx,)`` as returned by ``el[w, s, :]``.
+
+    Returns
+    -------
+    flat : numpy.ndarray, dtype int64
+        All events concatenated in pixel order.
+    offsets : numpy.ndarray, shape (nx+1,), dtype int64
+        ``flat[offsets[x]:offsets[x+1]]`` are the events for pixel *x*.
+    """
+    nx = len(row)
+    lengths = np.array([len(r) for r in row], dtype=np.int64)
+    offsets = np.empty(nx + 1, dtype=np.int64)
+    offsets[0] = 0
+    np.cumsum(lengths, out=offsets[1:])
+    flat = (
+        np.concatenate(row).astype(np.int64)
+        if offsets[-1] > 0
+        else np.empty(0, dtype=np.int64)
+    )
+    return flat, offsets
+
+
+@jit_ifnumba(cache=True)
+def _accumulate_peak_counts(
+    flat,
+    offsets,
+    nx,
+    clock_ratio,
+    nbr_samples,
+    mass_axis,
+    sorted_low,
+    sorted_high,
+    npeaks,
+    result_sx,
+):  # pragma: no cover
+    """
+    Accumulate per-pixel peak counts into *result_sx* (sorted-peak order).
+
+    Parameters
+    ----------
+    flat : int64 array
+        Flat event buffer from :func:`_flatten_event_row`.
+    offsets : int64 array, shape (nx+1,)
+        Pixel start/end indices into *flat*.
+    nx : int
+        Number of pixels in this row.
+    clock_ratio : int
+        TDC-to-ADC sample index divisor.
+    nbr_samples : int
+        Length of MassAxis; events outside [0, nbr_samples) are discarded.
+    mass_axis : float64 array
+        Calibrated mass (Da) for each ADC sample index.
+    sorted_low, sorted_high : float64 arrays, shape (npeaks,)
+        Integration window bounds in ascending mass order.
+    npeaks : int
+        Number of peaks.
+    result_sx : float32 array, shape (nx, npeaks)
+        Output accumulator, pre-zeroed by caller.  Indexed by sorted peak order.
+    """
+    for x in range(nx):
+        start = offsets[x]
+        end = offsets[x + 1]
+        for i in range(start, end):
+            sample = flat[i] // clock_ratio
+            if sample < 0 or sample >= nbr_samples:
+                continue
+            mass = mass_axis[sample]
+            bin_idx = np.searchsorted(sorted_low, mass, side="right") - 1
+            if bin_idx >= 0 and mass <= sorted_high[bin_idx]:
+                result_sx[x, bin_idx] += 1
+
+
+def compute_peak_data_from_eventlist(
+    event_list,
+    mass_axis,
+    nbr_samples,
+    clock_ratio,
+    normalization,
+    peak_table,
+    show_progressbar=True,
+):
+    """
+    Reconstruct the 4-D peak-integrated array from raw EventList data.
+
+    This replicates the processing performed by the Tofwerk proprietary
+    software when opening a raw file:
+
+    1. For each pixel's variable-length EventList (raw TDC timestamps),
+       convert timestamps to ADC sample indices by integer division with the
+       TDC-to-ADC clock ratio (``SampleInterval / ClockPeriod``, typically 64).
+    2. Look up the calibrated mass for each sample index via ``MassAxis``.
+    3. Count events that fall within each peak's integration window.
+    4. Divide by ``NbrWaveforms × NActiveChannels`` — the number of times each
+       physical ion is recorded in the EventList (once per ToF cycle per active
+       recording channel).
+
+    .. note::
+        If ``numba`` is installed, events are processed via a JIT-compiled loop
+        over a flat event buffer (no intermediate arrays per pixel).  Otherwise
+        falls back to a vectorised NumPy path that allocates per-pixel arrays but
+        is still significantly faster than a pure-Python triple loop.
+
+    Parameters
+    ----------
+    event_list : array-like
+        Ragged object array of shape ``(nwrites, nsegs, nx)`` of uint16 TDC
+        timestamps, one variable-length array per pixel.  Accepts h5py
+        variable-length datasets or numpy object arrays (as loaded by
+        ``signal="event_list"``).
+    mass_axis : numpy.ndarray
+        1-D array of shape ``(nbr_samples,)`` with calibrated mass (Da) for
+        each ADC sample index, from ``FullSpectra/MassAxis``.
+    nbr_samples : int
+        Length of ``mass_axis``; events outside ``[0, nbr_samples)`` are
+        discarded.
+    clock_ratio : int
+        TDC-to-ADC sample index divisor (``SampleInterval / ClockPeriod``).
+        Pass 1 if ``ClockPeriod`` is not available.
+    normalization : int
+        ``NbrWaveforms × NActiveChannels`` — the divisor used to convert raw
+        event counts to per-waveform ion counts.
+    peak_table : list of dict
+        Integration windows.  Each dict must have keys
+        ``lower_integration_limit`` and ``upper_integration_limit`` (Da).
+    show_progressbar : bool, default=True
+        Whether to show tqdm progress bars during reconstruction.
+
+    Returns
+    -------
+    numpy.ndarray
+        Shape ``(nwrites, nsegs, nx, npeaks)``, dtype float32.
+        Identical in structure to ``PeakData/PeakData`` in a pre-processed file.
+    """
+    el = event_list
+    mass_axis = np.asarray(mass_axis, dtype=np.float64)
+    _, peak_low, peak_high = _peak_table_from_list(peak_table)
+    npeaks = len(peak_table)
+    nwrites, nsegs, nx = el.shape
+
+    # Sort peak windows for O(log n) assignment via searchsorted
+    sort_idx = np.argsort(peak_low)
+    sorted_low = peak_low[sort_idx]
+    sorted_high = peak_high[sort_idx]
+    # Inverse permutation: maps sorted-peak index back to original peak order.
+    # Used to avoid numpy's advanced-indexing-on-LHS shape quirk where
+    # result[w, s, :, sort_idx] = arr yields shape (npeaks, nx) not (nx, npeaks).
+    inv_sort_idx = np.argsort(sort_idx)
+
+    result = np.zeros((nwrites, nsegs, nx, npeaks), dtype=np.float32)
+
+    _logger.info(
+        f"Reconstructing peak data from EventList "
+        f"({nwrites} depth slices × {nsegs} × {nx} pixels × {npeaks} peaks)."
+    )
+
+    # Prefer the numba JIT path when numba is installed: events are packed into a
+    # flat contiguous buffer per row and processed in a single compiled loop,
+    # avoiding per-pixel Python object overhead entirely.  Falls back to a
+    # vectorised NumPy path (one array allocation per pixel) when numba is absent.
+    try:
+        import numba  # noqa: F401
+
+        _use_numba = True
+    except ImportError:
+        _use_numba = False
+
+    try:
+        from tqdm.auto import tqdm as _tqdm
+    except ImportError:
+        _tqdm = None
+
+    def _make_pbar(n, desc="Reconstructing peak data"):
+        if _tqdm is not None and show_progressbar:
+            return _tqdm(total=n, desc=desc, unit=" slices")
+        return None
+
+    if _use_numba:
+        # numba path: flatten each (write, segment) row into a contiguous int64
+        # buffer and dispatch to the JIT-compiled _accumulate_peak_counts kernel.
+        # result_sx is reused across rows to avoid repeated allocation.
+        # Avoids ~6 intermediate per-pixel arrays; benchmarked as ~19× faster than a
+        # per-pixel HDF5 read loop on a 2.2GB dataset.
+        result_sx = np.zeros((nx, npeaks), dtype=np.float32)
+        pbar = _make_pbar(nwrites)
+        for w in range(nwrites):
+            for s in range(nsegs):
+                # event_list (el) shape is (depth, y, x); w=depth slice, s=y row, :=all x
+                # pixels. Each pixel holds variable-length ToF clock ticks, so we
+                # flatten into a contiguous int64 buffer (flat) + per-pixel
+                # slice boundaries (offsets) for the Numba kernel.
+                flat, offsets = _flatten_event_row(el[w, s, :])
+                result_sx[:] = 0
+                _accumulate_peak_counts(
+                    flat,
+                    offsets,
+                    nx,
+                    clock_ratio,
+                    nbr_samples,
+                    mass_axis,
+                    sorted_low,
+                    sorted_high,
+                    npeaks,
+                    result_sx,
+                )
+                result[w, s] = result_sx[:, inv_sort_idx]
+            if pbar is not None:
+                pbar.update(1)
+        if pbar is not None:
+            pbar.close()
+    else:
+        # NumPy fallback path: iterate over pixels explicitly, converting each
+        # pixel's variable-length event array to masses and binning with
+        # searchsorted + bincount.  Slower than the numba path but avoids any
+        # compiled dependency.  Reading per row rather than per pixel is still
+        # Benchmarked as ~3× faster than a per-pixel HDF5 read loop on a 2.2GB dataset.
+        pbar = _make_pbar(nwrites)
+        for w in range(nwrites):
+            for s in range(nsegs):
+                row = el[w, s, :]
+                for x in range(nx):
+                    events = row[x]
+                    if len(events) == 0:
+                        continue
+                    processed = events // clock_ratio
+                    valid = (processed >= 0) & (processed < nbr_samples)
+                    if not valid.any():
+                        continue
+                    event_masses = mass_axis[processed[valid]]
+                    bin_idx = (
+                        np.searchsorted(sorted_low, event_masses, side="right") - 1
+                    )
+                    in_peak = (
+                        (bin_idx >= 0)
+                        & (bin_idx < npeaks)
+                        & (event_masses <= sorted_high[bin_idx])
+                    )
+                    counts = np.bincount(bin_idx[in_peak], minlength=npeaks)
+                    result[w, s, x, sort_idx] = counts
+            if pbar is not None:
+                pbar.update(1)
+        if pbar is not None:
+            pbar.close()
+
+    pbar = _make_pbar(nwrites, desc="Normalising")
+    for w in range(nwrites):
+        result[w] /= normalization
+        if pbar is not None:
+            pbar.update(1)
+    if pbar is not None:
+        pbar.close()
+    return result
